remove old misc file and add 2D case
